Refactor: generalize scheduler
